+++ src/app/js/util/items-viewer.js
@@ -17,6 +17,7 @@
       controller: ItemsViewerController,
       scope: {
         columnDefs: '=',
+        someColumnDefs: '=',
         items:      '='
       }
     }
@@ -44,11 +45,39 @@
       ' title="TOOLTIP">{{COL_FIELD CUSTOM_FILTERS}}' +
       '</div></div>';
 
-    //console.table($scope.columnDefs);
+
+    var columnDefs;
+    if ($scope.columnDefs) {
+      columnDefs = $scope.columnDefs;
+    }
+    else {
+      if ($scope.someColumnDefs) {
+        columnDefs = _.clone($scope.someColumnDefs);
+        var collectedKeys = _.map(columnDefs, "field");
+        var max = Math.min($scope.items.length, 20);
+        for (var ii = 0; ii < max; ii++) {
+          var itemKeys = _.keys($scope.items[ii]);
+          _.each(itemKeys, function(itemKey) {
+            if (!_.contains(collectedKeys, itemKey)) {
+              collectedKeys.push(itemKey);
+              columnDefs.push({field: itemKey});
+            }
+          })
+        }
+      }
+    }
+    if (columnDefs) {
+      _.each(columnDefs, function(columnDef) {
+        if (columnDef.cellTemplate === undefined) {
+          columnDef.cellTemplate = mklinksOnlyExternal;
+        }
+      })
+    }
+    //console.table(columnDefs);
 
     $scope.gridOptions = {
       data: 'items',
-      columnDefs: $scope.columnDefs
+      columnDefs: columnDefs
       ,enableGridMenu: true
       ,showGridFooter: true
       ,enableFiltering: true
@@ -57,17 +86,17 @@
       }
     };
 
-    var vm = $scope.vm = {allowGrouping: false};
-    vm.allowGrouping = false;
-    $scope.$watch("vm.allowGrouping", function(allowGrouping) {
-      //console.debug("vm.allowGrouping=", allowGrouping);
-      _.each($scope.gridOptions.columnDefs, function(cd) {
-        cd.cellTemplate = allowGrouping ? groupingCellTemplate : mklinksOnlyExternal;
-      });
-      if (allowGrouping && $scope.gridApi && $scope.gridApi.treeBase) {
-        $scope.gridApi.treeBase.toggleRowTreeState($scope.gridApi.grid.renderContainers.body.visibleRowCache[0]);
-      }
-    });
+    //var vm = $scope.vm = {allowGrouping: false};
+    //vm.allowGrouping = false;
+    //$scope.$watch("vm.allowGrouping", function(allowGrouping) {
+    //  //console.debug("vm.allowGrouping=", allowGrouping);
+    //  _.each($scope.gridOptions.columnDefs, function(cd) {
+    //    cd.cellTemplate = allowGrouping ? groupingCellTemplate : mklinksOnlyExternal;
+    //  });
+    //  if (allowGrouping && $scope.gridApi && $scope.gridApi.treeBase) {
+    //    $scope.gridApi.treeBase.toggleRowTreeState($scope.gridApi.grid.renderContainers.body.visibleRowCache[0]);
+    //  }
+    //});
   }
 
 })();